refactor: live controls in preview
